// app/features/indicators/language/types.ts
export type IndicatorFill = {
  id: string
  title: string
  paneIndex: number
  points: Array<{ time: number; top: number; bottom: number; color: string }>
}

export type IndicatorBox = {
  id: string
  paneIndex: number
  leftTime: number
  rightTime?: number
  top: number
  bottom: number
  fillColor: string
  borderColor?: string
  borderWidth: number
  extendRight: boolean
}